Refactor code for building a Task object
Removed updateCard() and created function to build a Task that gets
called when updating a card and creating a new card.

Also fixed issue with else block creating a new Task to DB when it
should be skipped. Now will only create new entry if no match found
in DB.